+++ Assessment2/OLD DATASET/Lambda_Environment (LIVE) (DEPRECATED)/lambda_function.py
@@ -1,0 +1,10 @@
+from model import synchronizedCountryCovidData, sortedCovidData, sortedCountryData
+
+
+def lambda_handler(event, context):
+    print(event)
+    requestedData=synchronizedCountryCovidData
+    return {
+        'statusCode': 200,
+        'body': requestedData
+    }
